requirements.txt
google-genai
flask
PyPDF2
